### 2023-03-09 #5239 @editorialbot
[PRE REVIEW]: pyGeo: A geometry package for multidisciplinary design optimization
**Submitting author:** @hajdik (<a href="http://orcid.org/0000-0002-5103-7159">Hannah Hajdik</a>)
**Repository:** https://github.com/mdolab/pygeo
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.12.3
