### 2023-03-09 #5239 @danielskatz
👋 @hajdik - thanks for your submission.

re the suggested DOI, please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.  This appears to be a correct (new) DOI for this (older) report.

Also, note that your paper is almost twice as long as suggested for JOSS papers. Perhaps some of the details of the feature discussion can be moved into documentation, then that could be pointed to in this paper?
