+++ pygeo/mphys/mphys_dvgeo.py
@@ -71,10 +71,8 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         # compute the DVCon constraint values
         constraintfunc = dict()
         self.DVCon.evalFunctions(constraintfunc, includeLinear=True)
-        comm = self.comm
-        if comm.rank == 0:
-            for constraintname in constraintfunc:
-                outputs[constraintname] = constraintfunc[constraintname]
+        for constraintname in constraintfunc:
+            outputs[constraintname] = constraintfunc[constraintname]
 
         # we ran a compute so the inputs changed. update the dvcon jac
         # next time the jacvec product routine is called
@@ -212,66 +210,38 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
 
     def nom_addThicknessConstraints2D(self, name, leList, teList, nSpan=10, nChord=10):
         self.DVCon.addThicknessConstraints2D(leList, teList, nSpan, nChord, lower=1.0, name=name)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.ones((nSpan * nChord,)), shape=nSpan * nChord)
-        else:
-            self.add_output(name, distributed=True, shape=(0,))
+        self.add_output(name, distributed=False, val=np.ones((nSpan * nChord,)), shape=nSpan * nChord)
 
     def nom_addThicknessConstraints1D(self, name, ptList, nCon, axis):
         self.DVCon.addThicknessConstraints1D(ptList, nCon, axis, name=name)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.ones(nCon), shape=nCon)
-        else:
-            self.add_output(name, distributed=True, shape=(0))
+        self.add_output(name, distributed=False, val=np.ones(nCon), shape=nCon)
 
     def nom_addVolumeConstraint(self, name, leList, teList, nSpan=10, nChord=10):
         self.DVCon.addVolumeConstraint(leList, teList, nSpan=nSpan, nChord=nChord, name=name)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=1.0)
-        else:
-            self.add_output(name, distributed=True, shape=0)
+        self.add_output(name, distributed=False, val=1.0)
 
     def nom_add_LETEConstraint(self, name, volID, faceID, topID=None, childIdx=None):
         self.DVCon.addLeTeConstraints(volID, faceID, name=name, topID=topID, childIdx=childIdx)
         # how many are there?
         conobj = self.DVCon.linearCon[name]
         nCon = len(conobj.indSetA)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.zeros((nCon,)), shape=nCon)
-        else:
-            self.add_output(name, distributed=True, shape=0)
+        self.add_output(name, distributed=False, val=np.zeros((nCon,)), shape=nCon)
         return nCon
 
     def nom_addLERadiusConstraints(self, name, leList, nSpan, axis, chordDir):
         self.DVCon.addLERadiusConstraints(leList=leList, nSpan=nSpan, axis=axis, chordDir=chordDir, name=name)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.ones(nSpan), shape=nSpan)
-        else:
-            self.add_output(name, distributed=True, shape=0)
+        self.add_output(name, distributed=False, val=np.ones(nSpan), shape=nSpan)
 
     def nom_addCurvatureConstraint1D(self, name, start, end, nPts, axis, **kwargs):
         self.DVCon.addCurvatureConstraint1D(start=start, end=end, nPts=nPts, axis=axis, name=name, **kwargs)
-        comm = self.comm
[…]
