### 2023-03-10 1a15f29
Address flake8 rst docstring issues (#190)
* Address flake8 rst docstring issues

* typo
