+++ pygeo/constraints/DVCon.py
@@ -3139,6 +3139,8 @@ class DVConstraints:
         self, key, slope=1.0, name=None, start=0, stop=-1, config=None, childIdx=None, comp=None, DVGeoName="default"
     ):
         """
+        Add monotonic constraints to a given design variable.
+
         Parameters
         ----------
         key : str
