+++ paper/paper.bib
@@ -66,7 +66,6 @@
     title    = {{OpenMDAO}: An open-source framework for multidisciplinary design, analysis, and optimization},
     doi      = {10.1007/s00158-019-02211-z},
     journal  = {Structural and Multidisciplinary Optimization},
-    keywords = {openmdao},
     month    = {April},
     number   = {4},
     pages    = {1075--1104},
@@ -103,7 +102,6 @@
     title    = {Survey of Shape Parameterization Techniques for High-Fidelity Multidisciplinary Shape Optimization},
     doi      = {10.2514/2.1391},
     journal  = {{AIAA} Journal},
-    keywords = {geometry},
     number   = {5},
     pages    = {877--884},
     volume   = {39},
@@ -115,7 +113,6 @@
     title    = {{pyOptSparse:} A {Python} framework for large-scale constrained nonlinear optimization of sparse systems},
     doi      = {10.21105/joss.02564},
     journal  = {Journal of Open Source Software},
-    keywords = {ccavd},
     month    = {October},
     number   = {54},
     pages    = {2564},
@@ -128,7 +125,6 @@
     title     = {Aerostructural Wing Optimization for a Hydrogen Fuel Cell Aircraft},
     booktitle = {Proceedings of the AIAA SciTech Forum},
     doi       = {10.2514/6.2021-1132},
-    keywords  = {ank},
     month     = {January},
     year      = {2021}
 }
@@ -138,7 +134,6 @@
     title     = {Coupled Aeropropulsive Design Optimization of a Podded Electric Propulsor},
     booktitle = {AIAA Aviation Forum},
     doi       = {10.2514/6.2021-3032},
-    keywords  = {ank, aeroprop, openmdao, beststudentpaper},
     month     = {August},
     year      = {2021}
 }
@@ -176,7 +171,6 @@
     title    = {The Complex-Step Derivative Approximation},
     doi      = {10.1145/838250.838251},
     journal  = {ACM Transactions on Mathematical Software},
-    keywords = {complex step, sensitivities},
     month    = {September},
     number   = {3},
     pages    = {245--262},
@@ -190,7 +184,6 @@
     title    = {Component-based Geometry Manipulation for Aerodynamic Shape Optimization with Overset Meshes},
     doi      = {10.2514/1.J056550},
     journal  = {AIAA Journal},
-    keywords = {xsede},
     month    = {September},
     number   = {9},
     pages    = {3667--3679},
@@ -203,7 +196,6 @@
     title    = {A Surface Mesh Deformation Method Near Component Intersections for High-Fidelity Design Optimization},
     doi      = {10.1007/s00366-020-01247-w},
     journal  = {Engineering with Computers},
-    keywords = {ank, xsede},
     month    = {January},
     year     = {2021}
 }
@@ -213,7 +205,6 @@
     title     = {Aerodynamic Shape Optimization of a Supersonic Transport Considering Low-Speed Stability},
     booktitle = {AIAA SciTech Forum},
     doi       = {10.2514/6.2022-2177},
-    keywords  = {ank, scalos},
     month     = {January},
     year      = {2022}
 }
@@ -233,7 +224,6 @@
     title    = {Multipoint Aerodynamic Shape Optimization for Subsonic and Supersonic Regimes},
     doi      = {10.2514/1.C036216},
     journal  = {Journal of Aircraft},
-    keywords = {ank},
     month    = {May},
     number   = {3},
     pages    = {650--662},
@@ -257,7 +247,6 @@
[…]
