### 2023-03-13 ba05a17
implement geograd fail flag as analysis error
