+++ pygeo/constraints/areaConstraint.py
@@ -258,7 +258,7 @@ class TriangulatedSurfaceConstraint(GeometricConstraint):
         if self.perim_length > self.max_perim:
             failflag = True
             if self.comm.rank == 0:
-                print("Intersection length ", str(perim_length), " exceeds tol, returning fail flag")
+                print(f"Intersection length {self.perim_length} exceeds tol {self.max_perim}, returning fail flag")
         else:
             failflag = False
         return KS, perim_length, failflag
