### 2023-03-13 f8142c1
fix for new OM parallelism convention
