+++ pygeo/constraints/areaConstraint.py
@@ -1,5 +1,4 @@
 # External modules
-from mpi4py import MPI
 import numpy as np
 
 # Local modules
