### 2023-03-13 11f05db
eytan says remove default shape argument
