### 2023-03-14 5b7066e
Add triangulated surface constraint to MPhys wrapper (#192)
* fix comm

* add trisurfcon to mphys wrapper

* add missing options to constraints

* implement geograd fail flag as analysis error

* fix for new OM parallelism convention

* bye mpi

* add comm to DVCon tests

* eytan says remove default shape argument
