### 2023-03-14 0e6b356
Fixing Constraints for new OM Parallelism Convention (#191)
* Fixing constraints for new OM parallelism convention

* Fixing formatting
