+++ pygeo/mphys/mphys_dvgeo.py
@@ -72,16 +72,13 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         # compute the DVCon constraint values
         constraintfunc = dict()
         self.DVCon.evalFunctions(constraintfunc, includeLinear=True)
-        comm = self.comm
 
         for constraintname in constraintfunc:
             # if any constraint returned a fail flag throw an error to OpenMDAO
             # all constraints need the same fail flag, no <name_> prefix
             if constraintname == "fail":
                 raise AnalysisError("Analysis error in geometric constraints")
-
-            if comm.rank == 0:
-                outputs[constraintname] = constraintfunc[constraintname]
+            outputs[constraintname] = constraintfunc[constraintname]
 
         # we ran a compute so the inputs changed. update the dvcon jac
         # next time the jacvec product routine is called
@@ -219,66 +216,38 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
 
     def nom_addThicknessConstraints2D(self, name, leList, teList, nSpan=10, nChord=10):
         self.DVCon.addThicknessConstraints2D(leList, teList, nSpan, nChord, lower=1.0, name=name)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.ones((nSpan * nChord,)), shape=nSpan * nChord)
-        else:
-            self.add_output(name, distributed=True, shape=(0,))
+        self.add_output(name, distributed=False, val=np.ones((nSpan * nChord,)), shape=nSpan * nChord)
 
     def nom_addThicknessConstraints1D(self, name, ptList, nCon, axis, scaled=True):
         self.DVCon.addThicknessConstraints1D(ptList, nCon, axis, name=name, scaled=scaled)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.ones(nCon), shape=nCon)
-        else:
-            self.add_output(name, distributed=True, shape=(0))
+        self.add_output(name, distributed=False, val=np.ones(nCon), shape=nCon)
 
     def nom_addVolumeConstraint(self, name, leList, teList, nSpan=10, nChord=10, surfaceName="default"):
         self.DVCon.addVolumeConstraint(leList, teList, nSpan=nSpan, nChord=nChord, name=name, surfaceName=surfaceName)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=1.0)
-        else:
-            self.add_output(name, distributed=True, shape=0)
+        self.add_output(name, distributed=False, val=1.0)
 
     def nom_add_LETEConstraint(self, name, volID, faceID, topID=None, childIdx=None):
         self.DVCon.addLeTeConstraints(volID, faceID, name=name, topID=topID, childIdx=childIdx)
         # how many are there?
         conobj = self.DVCon.linearCon[name]
         nCon = len(conobj.indSetA)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.zeros((nCon,)), shape=nCon)
-        else:
-            self.add_output(name, distributed=True, shape=0)
+        self.add_output(name, distributed=False, val=np.zeros((nCon,)), shape=nCon)
         return nCon
 
     def nom_addLERadiusConstraints(self, name, leList, nSpan, axis, chordDir):
         self.DVCon.addLERadiusConstraints(leList=leList, nSpan=nSpan, axis=axis, chordDir=chordDir, name=name)
-        comm = self.comm
-        if comm.rank == 0:
-            self.add_output(name, distributed=True, val=np.ones(nSpan), shape=nSpan)
-        else:
-            self.add_output(name, distributed=True, shape=0)
[…]
