### 2023-03-20 #5239 @drvinceknight
@vissarion @HaoZeke @zhaowei0566 I am pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “pyGeo: A geometry package for multidisciplinary design optimization”. You can find more information at the top of this Github issue.
