### 2023-03-28 fbf6ad4
changed child bookkeeping to use a dict instead of list
