+++ pygeo/parameterization/DVGeo.py
@@ -118,7 +118,7 @@ class DVGeometry(BaseDVGeometry):
 
         # Flags to determine if this DVGeometry is a parent or child
         self.isChild = child
-        self.children = []
+        self.children = OrderedDict()
         self.iChild = None
         self.masks = None
         self.finalized = False
@@ -788,13 +788,13 @@ class DVGeometry(BaseDVGeometry):
             self.FFD._updateVolumeCoef()
 
         # Now embed into the children:
-        for child in self.children:
+        for child in self.children.values():
             child.addPointSet(points, ptName, origConfig, **kwargs)
 
         self.FFD.calcdPtdCoef(ptName)
         self.updated[ptName] = False
 
-    def addChild(self, childDVGeo):
+    def addChild(self, childDVGeo, childName=None):
         """Embed a child FFD into this object.
 
         An FFD child is a 'sub' FFD that is fully contained within
@@ -823,18 +823,22 @@ class DVGeometry(BaseDVGeometry):
         iChild = len(self.children)
         childDVGeo.iChild = iChild
 
-        self.FFD.attachPoints(childDVGeo.FFD.coef, "child%d_coef" % (iChild))
-        self.FFD.calcdPtdCoef("child%d_coef" % (iChild))
+        # check if a custom name is provided, if not, we will use the old naming scheme based on the iChild
+        if childName is None:
+            childName = f"child{iChild:d}"
+
+        self.FFD.attachPoints(childDVGeo.FFD.coef, f"{childName}_coef")
+        self.FFD.calcdPtdCoef(f"{childName}_coef")
 
         # We must finalize the Child here since we need the ref axis
         # coefficients
         if len(childDVGeo.axis) > 0:
             childDVGeo._finalizeAxis()
-            self.FFD.attachPoints(childDVGeo.refAxis.coef, "child%d_axis" % (iChild))
-            self.FFD.calcdPtdCoef("child%d_axis" % (iChild))
+            self.FFD.attachPoints(childDVGeo.refAxis.coef, f"{childName}_axis")
+            self.FFD.calcdPtdCoef(f"{childName}_axis")
 
         # Add the child to the parent and return
-        self.children.append(childDVGeo)
+        self.children[childName] = childDVGeo
 
     def addGlobalDV(self, dvName, value, func, lower=None, upper=None, scale=1.0, config=None):
         """
@@ -1598,7 +1602,7 @@ class DVGeometry(BaseDVGeometry):
 
         # Now call setValues on the children. This way the
         # variables will be set on the children
-        for child in self.children:
+        for child in self.children.values():
             child.setDesignVars(dvDict)
 
     def zeroJacobians(self, ptSetNames):
@@ -1644,7 +1648,7 @@ class DVGeometry(BaseDVGeometry):
         # Now call getValues on the children. This way the
         # returned dictionary will include the variables from
         # the children
-        for child in self.children:
+        for child in self.children.values():
             childdvDict = child.getValues()
             dvDict.update(childdvDict)
 
@@ -1848,17 +1852,17 @@ class DVGeometry(BaseDVGeometry):
             self.FFD.coef = self.origFFDCoef.copy()
             self._setInitialValues()
 
-            for iChild in range(len(self.children)):
-                if len(self.children[iChild].axis) > 0:
-                    self.children[iChild]._finalize()
-                    refaxis_ptSetName = "child%d_axis" % (iChild)
+            for childName, child in self.children.items():
+                if len(child.axis) > 0:
+                    child._finalize()
+                    refaxis_ptSetName = f"{childName}_axis"
                     if refaxis_ptSetName not in self.FFD.embeddedVolumes:
-                        self.FFD.attachPoints(self.children[iChild].refAxis.coef, refaxis_ptSetName)
[…]
