### 2023-03-28 183a89e
fix childidx
