### 2023-03-29 4aed9bb
added active child dvgeo handling
