### 2023-03-31 #5319 @editorialbot
[REVIEW]: pyGeo: A geometry package for multidisciplinary design optimization
**Submitting author:** @hajdik (<a href="http://orcid.org/0000-0002-5103-7159">Hannah Hajdik</a>)
**Repository:** https://github.com/mdolab/pygeo
**Branch with paper.md** (empty if default branch): main
**Version:** v1.13.0
