### 2023-04-03 a7d57d6
add other DVGeo and DVCon functions
