+++ pygeo/mphys/mphys_dvgeo.py
@@ -84,6 +84,10 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         # next time the jacvec product routine is called
         self.update_jac = True
 
+    """
+    Wrapper for DVGeo functions
+    """
+
     def nom_addChild(self, ffd_file):
         # can only add a child to a FFD DVGeo
         if self.geo_type != "ffd":
@@ -187,32 +191,40 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
 
         # add the DV to a normal DVGeo
         if childIdx is None:
-            nVal = self.DVGeo.addLocalSectionDV(
-                dvName=dvName,
-                secIndex=secIndex,
-                axis=axis,
-                pointSelect=pointSelect,
-                volList=volList,
-                orient0=orient0,
-                orient2=orient2,
-                config=config,
-            )
+            nVal = self.DVGeo.addLocalSectionDV(dvName, secIndex, axis, pointSelect, volList, orient0, orient2, config)
         # add the DV to a child DVGeo
         else:
             nVal = self.DVGeo.children[childIdx].addLocalSectionDV(
-                dvName=dvName,
-                secIndex=secIndex,
-                axis=axis,
-                pointSelect=pointSelect,
-                volList=volList,
-                orient0=orient0,
-                orient2=orient2,
-                config=config,
+                dvName,
+                secIndex,
+                axis,
+                pointSelect,
+                volList,
+                orient0,
+                orient2,
+                config,
             )
 
         # define the input
         self.add_input(dvName, distributed=False, shape=nVal)
         return nVal
+
+    def nom_getChildList(self):
+        return self.DVGeo.children
+
+    def nom_getLocalIndex(self, iVol, childIdx=None):
+        # this function is only for FFD-based DVGeo objects
+        if self.geo_type != "ffd":
+            raise RuntimeError(f"Only FFD-based DVGeo objects can use getLocalIndex(), not type:{self.geo_type}")
+
+        # get local index from the parent DVGeo
+        if childIdx is None:
+            lIndex = self.DVGeo.getLocalIndex(iVol)
+        # get local index from a child DVGeo
+        else:
+            lIndex = self.DVGeo.children[childIdx].getLocalIndex(iVol)
+
+        return lIndex
 
     def nom_addGeoCompositeDV(self, dvName, ptSetName=None, u=None, scale=None, **kwargs):
         # call the dvgeo object and add this dv
@@ -259,8 +271,56 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         if not isComposite:
             self.add_input(desmptr_name, distributed=False, shape=val.shape, val=val)
 
-    def nom_addThicknessConstraints2D(self, name, leList, teList, nSpan=10, nChord=10):
-        self.DVCon.addThicknessConstraints2D(leList, teList, nSpan, nChord, lower=1.0, name=name)
+    def nom_addRefAxis(self, childIdx=None, **kwargs):
+        # references axes are only needed in FFD-based DVGeo objects
+        if self.geo_type != "ffd":
+            raise RuntimeError(f"Only FFD-based DVGeo objects can use reference axes, not type:{self.geo_type}")
+
+        # we just pass this through
+        if childIdx is None:
+            return self.DVGeo.addRefAxis(**kwargs)
+        else:
+            return self.DVGeo.children[childIdx].addRefAxis(**kwargs)
+
+    def nom_writeRefAxis(self, fileName, childIdx=None):
+        # references axes are only needed in FFD-based DVGeo objects
+        if self.geo_type != "ffd":
[…]
