### 2023-04-04 05c5e1d
Fix shape handling in `nom_addGlobalDV`
