+++ pygeo/mphys/mphys_dvgeo.py
@@ -136,21 +136,56 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             self.nom_addPointSet(v, k)
 
     def nom_addGlobalDV(self, dvName, value, func, childIdx=None, isComposite=False):
+        """Add a global design variable to the DVGeo object. This is a wrapper for the DVGeo.addGlobalDV method.
+
+        _extended_summary_
+
+        Parameters
+        ----------
+        dvName : str
+            A unique name to be given to this design variable group
+
+        value : float, or iterable list of floats
+            The starting value(s) for the design variable. This
+            parameter may be a single variable or a numpy array
+            (or list) if the function requires more than one
+            variable. The number of variables is determined by the
+            rank (and if rank ==1, the length) of this parameter.
+
+        func : python function
+            The python function handle that will be used to apply the
+            design variable
+
+        childIdx : int, optional
+            The zero-based index of the child FFD, if this DV is for a child FFD.
+            The index is defined by the order in which you add the child FFD to the parent.
+            For example, the first child FFD has an index of 0, the second an index of 1, and so on.
+
+        isComposite : bool, optional
+            Whether this DV is to be included in the composite DVs, by default False
+
+        Raises
+        ------
+        RuntimeError
+            Raised if the underlying DVGeo object is not an FFD
+        """
         # global DVs are only added to FFD-based DVGeo objects
         if self.geo_type != "ffd":
             raise RuntimeError(f"Only FFD-based DVGeo objects can use global DVs, not type:{self.geo_type}")
+
+        # call the dvgeo object and add this dv
+        if childIdx is None:
+            self.DVGeo.addGlobalDV(dvName, value, func)
+            shape = self.DVGeo.DV_listGlobal[dvName].nVal
+        else:
+            self.DVGeo.children[childIdx].addGlobalDV(dvName, value, func)
+            shape = self.DVGeo.children[childIdx].DV_listGlobal[dvName].nVal
 
         # define the input
         # When composite DVs are used, input is not required for the default DVs. Now the composite DVs are
         # the actual DVs. So OpenMDAO don't need the default DVs as inputs.
         if not isComposite:
-            self.add_input(dvName, distributed=False, shape=len(value))
-
-        # call the dvgeo object and add this dv
-        if childIdx is None:
-            self.DVGeo.addGlobalDV(dvName, value, func)
-        else:
-            self.DVGeo.children[childIdx].addGlobalDV(dvName, value, func)
+            self.add_input(dvName, val=value, distributed=False, shape=shape)
 
     def nom_addLocalDV(self, dvName, axis="y", pointSelect=None, childIdx=None, isComposite=False):
         # local DVs are only added to FFD-based DVGeo objects
