### 2023-04-04 e5b7aa5
cleanup docstring
