+++ doc/requirements.txt
@@ -1,2 +1,3 @@
 numpydoc
 sphinx_mdolab_theme
+openmdao
