### 2023-04-11 569cde4
undo shape func related changes
