+++ doc/requirements.txt
@@ -1,3 +1,4 @@
 numpydoc
 sphinx_mdolab_theme
 openmdao
+mphys
