+++ doc/requirements.txt
@@ -1,4 +1,2 @@
 numpydoc
 sphinx_mdolab_theme
-openmdao
-mphys
