### 2023-04-21 f6a44b1
Fix shape handling in `nom_addGlobalDV` (#194)
* Fix shape handling in `nom_addGlobalDV`

* cleanup docstring

* Point to original DVGeo method in docstring

* Create mphys component docs page

* Add openmdao to docs requirements

* Update conf.py

* Update requirements.txt

* Update conf.py

* `black -l 120 .`

* minor edit

* Update requirements.txt

---------

Co-authored-by: Eirikur Jonsson <eirikurj@umich.edu>
Co-authored-by: Hannah Hajdik <hajdik@umich.edu>
