+++ doc/conf.py
@@ -30,7 +30,18 @@ extensions.extend(
 )
 
 # mock import for autodoc
-autodoc_mock_imports = ["numpy", "mpi4py", "scipy", "pyspline", "baseclasses", "pysurf", "prefoil", "pyOCSM", "openvsp"]
+autodoc_mock_imports = [
+    "numpy",
+    "mpi4py",
+    "scipy",
+    "pyspline",
+    "baseclasses",
+    "pysurf",
+    "prefoil",
+    "pyOCSM",
+    "openvsp",
+    "openmdao",
+]
 
 # This sets the bibtex bibliography file(s) to reference in the documentation
 bibtex_bibfiles = ["ref.bib"]
