### 2023-04-26 f40c2db
reduce tolerance for Mac Mini tests
