### 2023-04-26 be5086a
add multiple dvgeos to one geometry component
