+++ pygeo/mphys/mphys_dvgeo.py
@@ -11,48 +11,46 @@ from .. import DVConstraints, DVGeometry, DVGeometryESP, DVGeometryVSP
 # class that actually calls the dvgeometry methods
 class OM_DVGEOCOMP(om.ExplicitComponent):
     def initialize(self):
-        self.options.declare("file", default=None)
-        self.options.declare("type", default=None)
-        self.options.declare("options", default=None)
-        self.options.declare("name", default="default")
-        # self.options.declare("DVCon", default=None)
+        self.options.declare("DVGeoInfo", default=None)
 
     def setup(self):
-        self.geo_type = self.options["type"]
+        self.DVGeoInfo = self.options["DVGeoInfo"]
+        self.DVGeos = {}
 
-        # create the DVGeo object that does the computations
-        if self.geo_type == "ffd":
-            # we are doing an FFD-based DVGeo
-            if self.options["options"] is None:
-                ffd_options = {}
-            else:
-                ffd_options = self.options["options"]
+        print(self.DVGeoInfo)
+        for name, info in self.DVGeoInfo.items():
+            # create the DVGeo object that does the computations
+            if info["type"] == "ffd":
+                # we are doing an FFD-based DVGeo
+                # if info["options"] is None:
+                #     ffd_options = {}
+                # else:
+                #     ffd_options = info["options"]
 
-            self.DVGeo = DVGeometry(self.options["file"], name=self.options["name"], **ffd_options)
+                self.DVGeos.update({name: DVGeometry(info["file"], name=info["name"])})
 
-        elif self.geo_type == "vsp":
-            # we are doing a VSP-based DVGeo
-            if self.options["options"] is None:
-                vsp_options = {}
-            else:
-                vsp_options = self.options["options"]
+            elif info["type"] == "vsp":
+                # we are doing a VSP-based DVGeo
+                if info["options"] is None:
+                    vsp_options = {}
+                else:
+                    vsp_options = info["options"]
 
-            self.DVGeo = DVGeometryVSP(self.options["file"], comm=self.comm, **vsp_options)
+                self.DVGeo.update({name: DVGeometryVSP(info["file"], comm=self.comm, **vsp_options)})
 
-        elif self.geo_type == "esp":
-            # we are doing an ESP-based DVGeo
-            if self.options["options"] is None:
-                esp_options = {}
-            else:
-                esp_options = self.options["options"]
+            elif info["type"] == "esp":
+                # we are doing an ESP-based DVGeo
+                # if info["options"] is None:
+                #     esp_options = {}
+                # else:
+                #     esp_options = info["options"]
 
-            self.DVGeo = DVGeometryESP(self.options["file"], comm=self.comm, **esp_options)
+                self.DVGeos.update({name: DVGeometryESP(info["file"], comm=self.comm, name=info["name"])})
 
-        # if self.options["DVCon"] == None:
+
         self.DVCon = DVConstraints()
-        # else:
-            # self.DVCon = self.options["DVCon"]
-        self.DVCon.setDVGeo(self.DVGeo, name=self.options["name"])
+        for _, DVGeo in self.DVGeos.items():
+            self.DVCon.setDVGeo(DVGeo, name=DVGeo.name)
 
         self.omPtSetList = []
 
@@ -66,14 +64,15 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
[…]
