### 2023-04-30 #5319 @zhaowei0566
@danielskatz This is a very powerful package for the airplane outer model line design, which enables a high-fidelity MDAO study. I have followed this optimization framework for a while. I have no problem with the manuscript just a minor question about the claim given in Line 54,  “Design variables formulated from groupings of FFD control points often exhibit ill conditioning”, do you have any relevant reference to support this, or was it studied in Wu et al. 2022?

The manuscript has covered relevant information and meets the JOSS criteria,  https://joss.readthedocs.io/en/latest/review_criteria.html. Good work!
