### 2023-05-01 a6ae3b9
change extra flags test check to compiler
