+++ pygeo/parameterization/DVGeoVSP.py
@@ -32,14 +32,14 @@ except ImportError:
 
 vspOutOfDate = False
 if vspInstalled:
-    vsp_version_str = openvsp.GetVSPVersion()
-    words = vsp_version_str.split()
-    vsp_version = words[-1]
+    vspVersionStr = openvsp.GetVSPVersion()
+    words = vspVersionStr.split()
+    vspVersion = words[-1]
     # VSP is installed, but version too old
-    if Version(vsp_version) < Version("3.28.0"):
+    if Version(vspVersion) < Version("3.28.0"):
         vspOutOfDate = True
     # Prior to OpenVSP 3.33.0, the "s" parameter varried between [0, 0.5]
-    elif Version(vsp_version) < Version("3.33.0"):
+    elif Version(vspVersion) < Version("3.33.0"):
         SMAX = 0.5
     # After this version, the range was changed to [0, 1.0].
     else:  # Version(vsp_version) >= Version("3.33.0")
@@ -108,18 +108,18 @@ class DVGeometryVSP(DVGeoSketch):
         super().__init__(fileName=fileName, comm=comm, scale=scale, projTol=projTol)
 
         if hasattr(openvsp, "VSPVehicle"):
-            self.vsp_model = openvsp.VSPVehicle()
+            self.vspModel = openvsp.VSPVehicle()
         else:
-            self.vsp_model = openvsp
+            self.vspModel = openvsp
 
         self.exportComps = []
 
         # Clear the vsp model
-        self.vsp_model.ClearVSPModel()
+        self.vspModel.ClearVSPModel()
 
         t1 = time.time()
         # read the model
-        self.vsp_model.ReadVSPFile(fileName)
+        self.vspModel.ReadVSPFile(fileName)
         t2 = time.time()
         if self.comm.rank == 0:
             print("Loading the vsp model took:", (t2 - t1))
@@ -127,10 +127,10 @@ class DVGeometryVSP(DVGeoSketch):
         # List of all components returned from VSP. Note that this
         # order is important. It is the order that we use to map the
         # actual geom_id by using the geom_names
-        allComps = self.vsp_model.FindGeoms()
+        allComps = self.vspModel.FindGeoms()
         allNames = []
         for c in allComps:
-            allNames.append(self.vsp_model.GetContainerName(c))
+            allNames.append(self.vspModel.GetContainerName(c))
 
         if not comps:
             # no components specified, we use all
@@ -146,7 +146,7 @@ class DVGeometryVSP(DVGeoSketch):
         self.bbox = OrderedDict()
         self.bboxuv = self._getuv()
         for c in self.allComps:
-            self.compNames.append(self.vsp_model.GetContainerName(c))
+            self.compNames.append(self.vspModel.GetContainerName(c))
             self.bbox[c] = self._getBBox(c)
 
         # Now, we need to form our own quad meshes for fast projections
@@ -293,7 +293,7 @@ class DVGeometryVSP(DVGeoSketch):
             # is linear through the depth (i.e. in t)
             tg = 0.5
             # Now, find the closest volume projection
-            d, r[i], s[i], t[i] = self.vsp_model.FindRSTGuess(gid, 0, pnt, rg, sg, tg)
+            d, r[i], s[i], t[i] = self.vspModel.FindRSTGuess(gid, 0, pnt, rg, sg, tg)
             geom[i] = gind
 
             # if we dont have a good projection, try projecting again to surfaces
@@ -311,7 +311,7 @@ class DVGeometryVSP(DVGeoSketch):
                         and (self.bbox[gid][2, 0] < points[i, 2] < self.bbox[gid][2, 1])
                     ):
                         # project the point onto the VSP geometry
[…]
