### 2023-05-02 fd843d3
fix nom_add_discipline_coords
