### 2023-05-10 957ee93
bugfix to the clipping and retrained the test
