### 2023-05-11 f2f7a86
moved the eps definition to class init and updated comment
