### 2023-05-11 ba53e25
Add `nom_addProjectedAreaConstraint` and allow non-scaled area/volume constraints (#196)
* Add `nom_addProjectedAreaConstraint`, add `scaled` option to `nom_addVolumeConstraint`

* Fix docstring link

---------

Co-authored-by: Galen <68572680+gang525@users.noreply.github.com>
