### 2023-05-12 #5319 @hajdik
> just a minor question about the claim given in Line 54, “Design variables formulated from groupings of FFD control points often exhibit ill conditioning”, do you have any relevant reference to support this, or was it studied in Wu et al. 2022?

This was studied in Wu et al. 2022. I moved the reference to clarify this @zhaowei0566.
