### 2023-05-15 d7f6d4d
add getter for dvgeo
