### 2023-05-15 5a577ba
clean up my docstrings
