### 2023-05-15 7cb9ff2
extend to return a child DVGeo
