### 2023-05-15 52ebeaf
Updating DVGeoVSP to work with OpenVSP 3.33.0 (#198)
* Adding VSPVehicle class to DVGeoVSP

* Updating DVGeoVSP for VSP 3.33.0 backwards-incompatible change

* Re-order imports

* Adding import guard

* Fixing import test for vsp

* flake fix

* switch to camelCase

---------

Co-authored-by: Hannah Hajdik <hajdik@umich.edu>
