+++ pygeo/parameterization/DVGeoVSP.py
@@ -6,6 +6,7 @@ import time
 from baseclasses.utils import Error
 from mpi4py import MPI
 import numpy as np
+from packaging.version import Version
 from pyspline.utils import searchQuads
 
 # Local modules
@@ -28,13 +29,21 @@ except ImportError:
         openvsp = None
         vspInstalled = False
 
-# make sure volume projection api is available
-try:
-    openvsp.CompPntRST
 
-    vspOutOfDate = False
-except AttributeError:
-    vspOutOfDate = True
+vspOutOfDate = False
+if vspInstalled:
+    vspVersionStr = openvsp.GetVSPVersion()
+    words = vspVersionStr.split()
+    vspVersion = words[-1]
+    # VSP is installed, but version too old
+    if Version(vspVersion) < Version("3.28.0"):
+        vspOutOfDate = True
+    # Prior to OpenVSP 3.33.0, the "s" parameter varried between [0, 0.5]
+    elif Version(vspVersion) < Version("3.33.0"):
+        SMAX = 0.5
+    # After this version, the range was changed to [0, 1.0].
+    else:  # Version(vsp_version) >= Version("3.33.0")
+        SMAX = 1.0
 
 
 class DVGeometryVSP(DVGeoSketch):
@@ -98,15 +107,19 @@ class DVGeometryVSP(DVGeoSketch):
 
         super().__init__(fileName=fileName, comm=comm, scale=scale, projTol=projTol)
 
+        if hasattr(openvsp, "VSPVehicle"):
+            self.vspModel = openvsp.VSPVehicle()
+        else:
+            self.vspModel = openvsp
+
         self.exportComps = []
 
         # Clear the vsp model
-        openvsp.ClearVSPModel()
+        self.vspModel.ClearVSPModel()
 
         t1 = time.time()
         # read the model
-
-        openvsp.ReadVSPFile(fileName)
+        self.vspModel.ReadVSPFile(fileName)
         t2 = time.time()
         if self.comm.rank == 0:
             print("Loading the vsp model took:", (t2 - t1))
@@ -114,10 +127,10 @@ class DVGeometryVSP(DVGeoSketch):
         # List of all components returned from VSP. Note that this
         # order is important. It is the order that we use to map the
         # actual geom_id by using the geom_names
-        allComps = openvsp.FindGeoms()
+        allComps = self.vspModel.FindGeoms()
         allNames = []
         for c in allComps:
-            allNames.append(openvsp.GetContainerName(c))
+            allNames.append(self.vspModel.GetContainerName(c))
 
         if not comps:
             # no components specified, we use all
@@ -133,7 +146,7 @@ class DVGeometryVSP(DVGeoSketch):
         self.bbox = OrderedDict()
         self.bboxuv = self._getuv()
         for c in self.allComps:
-            self.compNames.append(openvsp.GetContainerName(c))
+            self.compNames.append(self.vspModel.GetContainerName(c))
             self.bbox[c] = self._getBBox(c)
 
         # Now, we need to form our own quad meshes for fast projections
@@ -270,17 +283,17 @@ class DVGeometryVSP(DVGeoSketch):
             rg = ug
             if vg < 0.5:
                 # This point is on the lower surface
-                sg = vg
+                sg = SMAX * 2.0 * vg
             else:
                 # This point is on the upper surface
-                sg = 1.0 - vg
+                sg = SMAX * 2.0 * (1.0 - vg)
             # tg = 0.5 places the initial guess in the middle of the upper and lower surfaces for the volume
             # Note: If the point we're looking for actually lies on the surface openvsp's
[…]
