### 2023-05-16 7c033c8
replace DVGeo name with object for trisurfcon
