+++ pygeo/constraints/DVCon.py
@@ -265,7 +265,7 @@ class DVConstraints:
         dvDict : dict
             Dictionary of design variables. The keys of the dictionary
             must correspond to the design variable names. Any
-            additional keys in the dfvdictionary are simply ignored.
+            additional keys in the dv dictionary are simply ignored.
         """
 
         # loop over the generated constraint objects and add the necessary
@@ -320,7 +320,7 @@ class DVConstraints:
             constraints. Normally this can be false since pyOptSparse
             does not need linear constraints to be returned.
         """
-
+        
         # loop over the generated constraints and evaluate their function values
         for conTypeKey in self.constraints:
             constraint = self.constraints[conTypeKey]
@@ -1372,7 +1372,7 @@ class DVConstraints:
         addToPyOpt=True,
     ):
         """
-        Add a single triangulated surface constraint to an aerosurface.
+        Add a single triangulated surface constraint to an aerosurface using Geograd.
         This constraint is designed to keep a general 'blob' of watertight
         geometry contained within an aerodynamic hull (e.g., a wing)
 
@@ -1383,19 +1383,19 @@ class DVConstraints:
             This should be the surface with the larger number of triangles.
             By default, it's the ADflow triangulated surface mesh.
 
-        DVGeo_1_name : str
-            The name of the DVGeo object to associate surface_1 to.
+        DVGeo1 : DVGeo or None
+            The DVGeo object to associate surface_1 to.
             If None, surface_1 will remain static during optimization.
-            By default, it's the 'default' DVGeo object
+            By default, it's the 'default' DVGeo object.
 
         surface_2_name : str
             The name of the second triangulated surface to constrain.
             This should be the surface with the smaller number of triangles.
 
-        DVGeo_2_name : str
-            The name of the DVGeo object to associate surface_2 to.
+        DVGeo2 : DVGeo or None
+            The DVGeo object to associate surface_2 to.
             If None, surface_2 will remain static during optimization.
-            By default, it's the 'default' DVGeo object
+            By default, it's the 'default' DVGeo object.
 
         rho : float
             The rho factor of the KS function of min distance.
@@ -1419,8 +1419,8 @@ class DVConstraints:
              multiple DVCon objects and the constraint names need to
              be distinguished **OR** you are using this
              computation for something other than a direct constraint
-             in pyOpt, i.e. it is required for a subsequent
-             computation.
+             in pyOpt, i.e. it is required for a subsequent computation.
+             The MPhys wrapper sets this name for tracking in OpenMDAO.
 
         scale : float
             This is the optimization scaling of the
@@ -1437,16 +1437,8 @@ class DVConstraints:
             addToPyOpt=False, the lower, upper and scale variables are
             meaningless
         """
-        # if DVGeo1 is not None:
-        #     self._checkDVGeo(DVGeo1.name)
-        # else:
-        #     DVGeo1 = None
-        # if DVGeo2 is not None:
-        #     self._checkDVGeo(DVGeo2.name)
-        # else:
-        #     DVGeo2 = None
-        # if DVGeo1 is None and DVGeo2 is None:
-        #     raise ValueError("At least one DVGeo object must be specified")
+        if DVGeo1 is None and DVGeo2 is None:
+            raise ValueError("At least one DVGeo object must be specified")
 
         typeName = "triSurfCon"
         if typeName not in self.constraints:
