### 2023-05-16 89a26cc
checks for DVGeo input dictionary
