+++ pygeo/parameterization/DVGeoESP.py
@@ -118,7 +118,7 @@ class DVGeometryESP(DVGeoSketch):
         exclude_edge_projections=False,
         ulimits=None,
         vlimits=None,
-        name=None
+        name=None,
     ):
         if not ocsmImported:
             raise ImportError("OCSM and pyOCSM must be installed to use DVGeometryESP.")
@@ -126,9 +126,7 @@ class DVGeometryESP(DVGeoSketch):
             print("Initializing DVGeometryESP")
             t0 = time.time()
 
-        super().__init__(fileName=fileName, comm=comm, scale=scale, projTol=projTol, name=name)
-
-        self.geoType = "esp"
+        super().__init__(fileName=fileName, comm=comm, scale=scale, projTol=projTol, name=name, geoType="esp")
 
         self.maxproc = maxproc
         self.esp = True
