+++ pygeo/parameterization/DVGeoVSP.py
@@ -96,7 +96,7 @@ class DVGeometryVSP(DVGeoSketch):
             print("Initializing DVGeometryVSP")
             t0 = time.time()
 
-        super().__init__(fileName=fileName, comm=comm, scale=scale, projTol=projTol)
+        super().__init__(fileName=fileName, comm=comm, scale=scale, projTol=projTol, geoType="vsp")
 
         self.exportComps = []
 
