### 2023-05-17 9cfba34
minor change to how the bounds are enforced
