+++ pygeo/mphys/mphys_dvgeo.py
@@ -135,6 +135,45 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         for k, v in point_dict.items():
             self.nom_addPointSet(v, k)
 
+    def nom_getDVGeo(self, childIdx=None):
+        """
+        Gets the DVGeometry object held in the geometry component so DVGeo methods can be called directly on it
+
+        Parameters
+        ----------
+        childIdx : int, optional
+            The zero-based index of the child FFD, you want a child DVGeo returned
+
+        Returns
+        -------
+        self.DVGeo, DVGeometry object
+            DVGeometry object held by this geometry component
+
+        """
+
+        # return the top level DVGeo
+        if childIdx is None:
+            return self.DVGeo
+
+        # return a child DVGeo
+        else:
+            return self.DVGeo.children[childIdx]
+
+    def nom_getDVCon(self):
+        """
+        Gets the DVConstraints object held in the geometry component so DVCon methods can be called directly on it
+
+        Returns
+        -------
+        self.DVCon, DVConstraints object
+            DVConstraints object held by this geometry component
+        """
+        return self.DVCon
+
+    """
+    Wrapper for DVGeo functions
+    """
+
     def nom_addGlobalDV(self, dvName, value, func, childIdx=None, isComposite=False):
         """Add a global design variable to the DVGeo object. This is a wrapper for the DVGeo.addGlobalDV method.
 
@@ -197,6 +236,80 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             self.add_input(dvName, distributed=False, shape=nVal)
         return nVal
 
+    def nom_addLocalSectionDV(
+        self,
+        dvName,
+        secIndex,
+        childIdx=None,
+        axis=1,
+        pointSelect=None,
+        volList=None,
+        orient0=None,
+        orient2="svd",
+        config=None,
+    ):
+        """
+        Add one or more section local design variables to the DVGeometry object
+        Wrapper for :meth:`addLocalSectionDV <.DVGeometry.addLocalSectionDV>`
+        Input parameters are identical to those in wrapped function unless otherwise specified
+
+        Parameters
+        ----------
+        dvName : str
+            Name to give this design variable
+        secIndex : char or list of chars
+            See wrapped
+        childIdx : int, optional
+            The zero-based index of the child FFD, if this DV is for a child FFD
+            The index is defined by the order in which you add the child FFD to the parent
+            For example, the first child FFD has an index of 0, the second an index of 1, and so on
+        axis : int, optional
+            See wrapped
+        pointSelect : pointSelect object, optional
+            See wrapped
+        volList : list, optional
+            See wrapped
+        orient0 : orientation, optional
+            See wrapped
+        orient2 : str, optional
+            See wrapped
+        config : str or list, optional
+            See wrapped
+
+        Returns
+        -------
+        nVal, int
+            number of local section DVs
+
+        Raises
+        ------
+        RuntimeError
+            Raised if the underlying DVGeo parameterization is not FFD-based
+        """
+        # local DVs are only added to FFD-based DVGeo objects
+        if self.geo_type != "ffd":
+            raise RuntimeError(f"Only FFD-based DVGeo objects can use local DVs, not type:{self.geo_type}")
+
+        # add the DV to a normal DVGeo
+        if childIdx is None:
+            nVal = self.DVGeo.addLocalSectionDV(dvName, secIndex, axis, pointSelect, volList, orient0, orient2, config)
[…]
