+++ paper/cst_example.py
@@ -1,4 +1,5 @@
 import os
+
 import matplotlib.pyplot as plt
 import numpy as np
 from pygeo.parameterization.DVGeoCST import DVGeometryCST
