+++ paper/cst_example.py
@@ -2,6 +2,7 @@ import os
 
 import matplotlib.pyplot as plt
 import numpy as np
+
 from pygeo.parameterization.DVGeoCST import DVGeometryCST
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
