+++ paper/Global-FFD-DV-Demo.py
@@ -1,7 +1,8 @@
+# Standard Python modules
 import copy
 
+# External modules
 import numpy as np
-
 from pygeo import DVGeometry
 
 FFDFile = "ffd.xyz"
