### 2023-05-24 a8b22e3
added test for shape function dvs
