+++ paper/paper.md
@@ -60,15 +60,15 @@ pyGeo computes derivatives for all parameterization methods and constraints, fac
 # Features
 
 ## Integrations
-pyGeo is the geometry manipulation engine within the MDO of Aircraft Configurations at High Fidelity (MACH) framework [@Kenway2014a; @Kenway2014c], which specializes in high-fidelity aerostructural optimization.
-pyGeo is also integrated into MPhys[^1], a more general framework for high-fidelity multiphysics problems built with OpenMDAO [@Gray2019a].
-Both MACH and MPhys use pyOptSparse [@Wu2020a] to interface with optimization algorithms.
+pyGeo is the geometry manipulation engine within the MDO of Aircraft Configurations at High Fidelity (MACH) framework [@Kenway2014a; @Kenway2014c], specializing in high-fidelity aerostructural optimization.
+pyGeo can be used as a stand-alone package and is integrated into MPhys[^1], a more general framework for high-fidelity multiphysics problems built with OpenMDAO [@Gray2019a].
+MACH and MPhys use pyOptSparse [@Wu2020a] to interface with optimization algorithms.
 
 [^1]: \url{https://github.com/OpenMDAO/mphys}
 
-pyGeo's interface for design variables and constraints is independent of which solvers are accessing the geometry.
-This means that pyGeo geometries can interact with different types of solvers, such as structures and aerodynamics, in the same way.
-This also allows direct comparison of the behavior or performance of two different solvers within the same discipline using the same geometry parameterization for each, such as two different flow solvers [@Adler2022c].
+pyGeo's interface for design variables and constraints is independent of which disciplinary models access the geometry.
+This means that pyGeo can interact with different disciplines ( such as structures and aerodynamics) in the same way.
+This also facilitates a direct comparison of the behavior or performance of two alternative models for a discipline using the same geometry parameterization [@Adler2022c].
 
 ## Geometry Parameterization with pyGeo
 pyGeo contains several options for parameterizing geometry: variations on the FFD method, interfaces to external parametric modeling tools, and an analytic parameterization.
@@ -78,7 +78,7 @@ The choice of parameterization depends on the user's experience, the geometry de
 ### Free-form Deformation
 The free-form deformation (FFD) method [@Sederberg1986] is one of the most popular three-dimensional geometry parameterization approaches [@Zhang2018a].
 This approach embeds the entire reference geometry in a parameterized volume. 
-The set of control points that determine the shape of the volume are displaced to manipulate the points inside. 
+The set of control points that determine the shape of the volume is displaced to manipulate the points inside. 
 The user can have a high degree of control over the geometry by selecting different control point densities and locations.
 
 Individual control points can be moved to obtain local shape modifications.
@@ -88,8 +88,8 @@ These are referred to as _global_ design variables in pyGeo.
 For example, wing twist variables can be defined as rotations of the control points about a reference axis that runs along the wing. 
 \autoref{fig:FFD_DV} shows a few common planform design variables for an aircraft wing.
 
-Design variables formulated from groupings of FFD control points often exhibit ill conditioning [@Wu2022b]. 
-To alleviate this, a parameterization based on singular value decomposition is also possible within pyGeo.
+Design variables formulated from groupings of FFD control points often exhibit ill-conditioning. 
+A parameterization based on singular value decomposition is also possible within pyGeo to alleviate this issue [@Wu2022b].
 
 ![Examples of common wing planform design variables.\label{fig:FFD_DV}](ffd_dvs.pdf)
 
@@ -97,7 +97,7 @@ In addition to the basic FFD implementation, pyGeo offers two additional feature
 
 #### Hierarchical FFD
[…]
