### 2023-06-05 adf7377
update docstrings
