### 2023-06-08 f049680
citation and acronym fixes
