### 2023-06-09 2827954
Add JOSS paper (#203)
* added initial joss paper

* added GHA

* added some basic summary

* start parametric tools section

* actually commit the paper

* mystery paragraph

* constraints

* added FFD section

* constraints

* whitespace

* format fixes

* cst

* fix constraints

* wording

* start examples

* started statement of need

* added some comments

* clarify esp/vsp need

* Updated DVGeoCST description

* esp/vsp process

* expanded FFD a bit

* added some short text on parallelism and gradients

* added DVGeoMulti description and figure

* edited ffd derivatives

* comments and ed
[…]
