+++ paper/paper.bib
@@ -1,0 +1,368 @@
+@inproceedings{Kenway2010b,
+    author    = {Gaetan K.W. Kenway and Graeme. J. Kennedy and Joaquim R. R. A. Martins},
+    title     = {A {CAD}-Free Approach to High-Fidelity Aerostructural Optimization},
+    address   = {Fort Worth, TX},
+    booktitle = {Proceedings of the 13th AIAA/ISSMO Multidisciplinary Analysis Optimization Conference},
+    doi       = {10.2514/6.2010-9231},
+    month     = {September},
+    number    = {AIAA 2010-9231},
+    year      = {2010}
+}
+
+
+@inbook{McDonald2022a,
+    author    = {Robert A. McDonald and James R. Gloudemans},
+    title     = {Open Vehicle Sketch Pad: An Open Source Parametric Geometry and Analysis Tool for Conceptual Aircraft Design},
+    booktitle = {AIAA SCITECH 2022 Forum},
+    doi       = {10.2514/6.2022-0004},
+    month     = {January},
+    publisher = {American Institute of Aeronautics and Astronautics},
+    year      = {2022}
+}
+
+
+@inproceedings{Haimes2013a,
+    author    = {Haimes, Robert and Dannenhoffer, John},
+    title     = {The Engineering Sketch Pad: A Solid-Modeling, Feature-Based, Web-Enabled System for Building Parametric Geometry},
+    booktitle = {21st {AIAA} {Computational} {Fluid} {Dynamics} {Conference}},
+    doi       = {10.2514/6.2013-3073},
+    month     = {June},
+    publisher = {American Institute of Aeronautics and Astronautics},
+    series    = {Fluid Dynamics and Co-located Conferences},
+    year      = {2013}
+}
+
+@article{Sederberg1986,
+    author     = {Sederberg, Thomas W. and Parry, Scott R.},
+    title      = {Free-form Deformation of Solid Geometric Models},
+    address    = {New York, NY, USA},
+    doi        = {10.1145/15886.15903},
+    issn       = {0097-8930},
+    issue_date = {Aug. 1986},
+    journal    = {SIGGRAPH Comput. Graph.},
+    month      = {August},
+    number     = {4},
+    pages      = {151--160},
+    publisher  = {ACM},
+    volume     = {20},
+    year       = {1986}
+}
+
+@article{Kulfan2008,
+    author    = {B. M. Kulfan},
+    title     = {Universal Parametric Geometry Representation Method},
+    doi       = {10.2514/1.29958},
+    journal   = {Journal of Aircraft},
+    month     = {January},
+    number    = {1},
+    pages     = {142--158},
+    publisher = {American Institute of Aeronautics and Astronautics},
+    volume    = {45},
+    year      = {2008}
+}
+
+@article{Gray2019a,
+    author   = {Justin S. Gray and John T. Hwang and Joaquim R. R. A. Martins and Kenneth T. Moore and Bret A. Naylor},
+    title    = {{OpenMDAO}: An open-source framework for multidisciplinary design, analysis, and optimization},
+    doi      = {10.1007/s00158-019-02211-z},
+    journal  = {Structural and Multidisciplinary Optimization},
+    month    = {April},
+    number   = {4},
+    pages    = {1075--1104},
+    volume   = {59},
+    year     = {2019}
+}
+
+@article{Kenway2014c,
+    author  = {Gaetan K. W. Kenway and Joaquim R. R. A. Martins},
+    title   = {Multipoint High-Fidelity Aerostructural Optimization of a Transport Aircraft Configuration},
+    doi     = {10.2514/1.C032150},
[…]
