+++ paper/paper.md
@@ -1,0 +1,201 @@
+---
+title: 'pyGeo: A geometry package for multidisciplinary design optimization'
+tags:
+  - geometry
+  - Python
+authors:
+  - name: Hannah M. Hajdik
+    orcid: 0000-0002-5103-7159
+    affiliation: 1
+  - name: Anil Yildirim
+    orcid: 0000-0002-1814-9191
+    affiliation: 1
+  - name: Neil Wu
+    orcid: 0000-0001-8856-9661
+    affiliation: 1
+  - name: Benjamin J. Brelje
+    orcid: 0000-0002-9819-3028
+    affiliation: 1
+  - name: Sabet Seraj
+    orcid: 0000-0002-7364-0071
+    affiliation: 1
+  - name: Marco Mangano
+    orcid: 0000-0001-8495-3578
+    affiliation: 1
+  - name: Joshua L. Anibal
+    orcid: 0000-0002-7795-2523
+    affiliation: 1
+  - name: Eirikur Jonsson
+    orcid: 0000-0002-5166-3889
+    affiliation: 1
+  - name: Eytan J. Adler
+    orcid: 0000-0002-8716-1805
+    affiliation: 1
+  - name: Charles A. Mader
+    orcid: 0000-0002-2744-1151
+    affiliation: 1
+  - name: Gaetan K. W. Kenway
+    orcid: 0000-0003-1352-5458
+    affiliation: 1
+  - name: Joaquim R. R. A. Martins
+    orcid: 0000-0003-2143-1478
+    affiliation: 1
+affiliations:
+ - name: Department of Aerospace Engineering, University of Michigan
+   index: 1
+date: October 18, 2022
+bibliography: paper.bib
+header-includes: \usepackage{subcaption}
+---
+
+# Summary
+Geometry parameterization is a key challenge in shape optimization. 
+Parameterizations must accurately capture the design intent and perform well in optimization.
+In multidisciplinary design optimization (MDO), the parameterization must additionally represent the shape consistently across each discipline. 
+
+pyGeo is a geometry package for three-dimensional shape manipulation tailored for aerodynamic and multidisciplinary design optimization.
+It provides several methods for geometry parameterization, geometric constraints, and utility functions for geometry manipulation.
+pyGeo computes derivatives for all parameterization methods and constraints, facilitating efficient gradient-based optimization.
+
+# Features
+
+## Integrations
+pyGeo is the geometry manipulation engine within the MDO of Aircraft Configurations at High Fidelity (MACH) framework [@Kenway2014a; @Kenway2014c], specializing in high-fidelity aerostructural optimization.
+pyGeo can be used as a stand-alone package and is integrated into MPhys[^1], a more general framework for high-fidelity multiphysics problems built with OpenMDAO [@Gray2019a].
+MACH and MPhys use pyOptSparse [@Wu2020a] to interface with optimization algorithms.
+
+[^1]: \url{https://github.com/OpenMDAO/mphys}
+
+pyGeo's interface for design variables and constraints is independent of which disciplinary models access the geometry.
+This means that pyGeo can interact with different disciplines (such as structures and aerodynamics) in the same way.
+This also facilitates a direct comparison of the behavior or performance of two alternative models for a discipline using the same geometry parameterization [@Adler2022c].
+
+## Geometry Parameterization with pyGeo
+pyGeo contains several options for parameterizing geometry: variations on the free-form deformation (FFD) method, interfaces to external parametric modeling tools, and an analytic parameterization.
+Because each parameterization method uses a common interface for interacting with the rest of the MACH framework, any surface parameterization can be used in place of another within an optimization setup [@Hajdik2023a].
[…]
