### 2023-06-12 #5319 @nwu63
Hi @drvinceknight, we have released a version which is archived on zenodo.
version number: `v1.13.0`
DOI: `10.5281/zenodo.8027706`

The paper branch has also changed to `main` now that we have merged mdolab/pygeo#203.
Please let us know if there's anything else you need.
