### 2023-06-26 7a59058
Minimal implementation to check compatibility
