+++ pygeo/mphys/mphys_dvgeo.py
@@ -5,7 +5,7 @@ import openmdao.api as om
 from openmdao.api import AnalysisError
 
 # Local modules
-from .. import DVConstraints, DVGeometry, DVGeometryESP, DVGeometryVSP
+from .. import DVConstraints, DVGeometry, DVGeometryESP, DVGeometryVSP, DVGeometryMulti
 
 
 # class that actually calls the dvgeometry methods
@@ -45,6 +45,10 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
                 esp_options = self.options["options"]
 
             self.DVGeo = DVGeometryESP(self.options["file"], comm=self.comm, **esp_options)
+
+        elif self.geo_type == "multi":
+            # we are doing a multi-DVGeo-based DVGeo
+            self.DVGeo = DVGeometryMulti()
 
         self.DVCon = DVConstraints()
         self.DVCon.setDVGeo(self.DVGeo)
@@ -100,7 +104,26 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             if pointSet not in self.DVGeo.children[-1].points:
                 self.DVGeo.children[-1].addPointSet(self.DVGeo.points[pointSet], pointSet)
 
-    def nom_add_discipline_coords(self, discipline, points=None):
+    def nom_addComponent(self, comp, ffd_file, triMesh, scale=1.0):
+        if self.geo_type != "multi":
+            raise RuntimeError(
+                f"Only multi-based DVGeo objects can have components added to them, not type:{self.geo_type}"
+            )
+
+        # Add component
+        DVGeo = DVGeometry(ffd_file)
+        self.DVGeo.addComponent(comp=comp, DVGeo=DVGeo, triMesh=triMesh, scale=scale)
+
+    def nom_addIntersection(self, compA, compB, dStarA, dStarB, featureCurves, project, includeCurves, curveEpsDict, anisotropy):#, trackSurfaces):
+        if self.geo_type != "multi":
+            raise RuntimeError(
+                f"Only multi-based DVGeo objects can have intersections added to them, not type:{self.geo_type}"
+            )
+
+        # Add intersection
+        self.DVGeo.addIntersection(compA=compA, compB=compB, dStarA=dStarA, dStarB=dStarB, featureCurves=featureCurves, project=project, includeCurves=includeCurves, curveEpsDict=curveEpsDict, anisotropy=anisotropy)#, trackSurfaces=trackSurfaces)
+
+    def nom_add_discipline_coords(self, discipline, points=None, applyIC=False):
         # TODO remove one of these methods to keep only one method to add pointsets
 
         if points is None:
@@ -110,13 +133,16 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
 
         else:
             # we are provided with points. we can do the full initialization now
-            self.nom_addPointSet(points, "x_%s0" % discipline, add_output=False)
+            self.nom_addPointSet(points, "x_%s0" % discipline, add_output=False, applyIC=applyIC)
             self.add_input("x_%s_in" % discipline, distributed=True, val=points.flatten())
             self.add_output("x_%s0" % discipline, distributed=True, val=points.flatten())
 
-    def nom_addPointSet(self, points, ptName, add_output=True, **kwargs):
+    def nom_addPointSet(self, points, ptName, add_output=True, applyIC=False):
         # add the points to the dvgeo object
-        self.DVGeo.addPointSet(points.reshape(len(points) // 3, 3), ptName, **kwargs)
+        if self.geo_type == "multi":
+            self.DVGeo.addPointSet(points.reshape(len(points) // 3, 3), ptName, applyIC=applyIC)
+        else:
+            self.DVGeo.addPointSet(points.reshape(len(points) // 3, 3), ptName)
[…]
