### 2023-06-26 9625192
Fixing incorrect logical
