+++ pygeo/parameterization/DVGeoMulti.py
@@ -160,6 +160,7 @@ class DVGeometryMulti:
         project=False,
         marchDir=1,
         includeCurves=False,
+        slidingCurves=[],
         intDir=None,
         curveEpsDict=None,
         trackSurfaces=None,
@@ -208,6 +209,10 @@ class DVGeometryMulti:
 
         includeCurves : bool, optional
             Flag to specify whether to include features curves in the inverse-distance deformation.
+
+        slidingCurves : list, optional
+            The list of curves to project to, but on which the mesh nodes are not frozen in their initial positions.
+            This allows the mesh nodes to slide along the feature curve.
 
         intDir : int, optional
             If there are multiple intersection curves, this specifies which curve to choose.
@@ -266,6 +271,7 @@ class DVGeometryMulti:
                 project,
                 marchDir,
                 includeCurves,
+                slidingCurves,
                 intDir,
                 curveEpsDict,
                 trackSurfaces,
@@ -957,6 +963,7 @@ class CompIntersection:
         project,
         marchDir,
         includeCurves,
+        slidingCurves,
         intDir,
         curveEpsDict,
         trackSurfaces,
@@ -1146,6 +1153,10 @@ class CompIntersection:
         # flag to include feature curves in ID-warping
         self.incCurves = includeCurves
 
+        # list of curves that allow nodes to slide on them. we only use these for the projection step,
+        # but they are not included in the first line based IDWarp
+        self.slidingCurves = slidingCurves
+
         # direction to pick if we have multiple intersection curves
         self.intDir = intDir
 
@@ -1331,7 +1342,7 @@ class CompIntersection:
                     elemIDs[:] = elemIDs + 1
                     # (we need to do this separetely because Fortran will actively change elemIDs contents.
                     self.curveSearchAPI.mindistancecurve(
-                        ptsToCurves.T, self.seam0.T, self.seamConn.T + 1, xyzProj.T, tanProj.T, dist2, elemIDs
+                        ptsToCurves.T, self.seam0.T, self.seamConnFull.T + 1, xyzProj.T, tanProj.T, dist2, elemIDs
                     )
 
                     # Adjust indices back to Python standards
@@ -1441,7 +1452,7 @@ class CompIntersection:
         # we use the initial seam coordinates here
         coor = self.seam0
         # bar connectivity for the remeshed elements
-        conn = self.seamConn
+        conn = self.seamConnWarp
         # deltas for each point (nNode, 3) in size
         if self.seam.shape == self.seam0.shape:
             dr = self.seam - self.seam0
@@ -1548,7 +1559,7 @@ class CompIntersection:
         # we use the initial seam coordinates here
         coor = self.seam0
         # bar connectivity for the remeshed elements
-        conn = self.seamConn
+        conn = self.seamConnWarp
 
         # Get the two end points for the line elements
         r0 = coor[conn[:, 0]]
@@ -1685,7 +1696,7 @@ class CompIntersection:
             # conn of the current curve
             seamBeg = self.seamBeg[curveName]
             seamEnd = self.seamEnd[curveName]
-            curveConn = self.seamConn[seamBeg:seamEnd]
+            curveConn = self.seamConnFull[seamBeg:seamEnd]
 
             # Project these to the combined curves using pySurf
             # Get number of points
@@ -2692,7 +2703,8 @@ class CompIntersection:
                 self.distFeature = {}
 
             remeshedCurves = np.zeros((0, 3), dtype=self.dtype)
-            remeshedCurveConn = np.zeros((0, 2), dtype="int32")
+            remeshedCurveConnFull = np.zeros((0, 2), dtype="int32")
+            remeshedCurveConnWarp = np.zeros((0, 2), dtype="int32")
 
[…]
