### 2023-07-07 #5319 @hajdik
paper.bib has been updated to remove the duplicate citation.
