### 2023-07-09 27cfce4
compute jacvec product
