+++ pygeo/mphys/mphys_dvgeo.py
@@ -97,7 +97,6 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
                 # update this pointset and write it as output
                 outputs[ptName] = DVGeo.update(ptName).flatten()
 
-
     def compute(self, inputs, outputs):
         # check for inputs that have been added but the points have not been added to dvgeo
         for var in inputs.keys():
@@ -130,7 +129,13 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         # next time the jacvec product routine is called
         self.update_jac = True
 
-    def nom_addChild(self, DVGeo, ffd_file):
+    def nom_addChild(self, ffd_file, DVGeoName=None):
+        # if we have multiple DVGeos use the one specified by name
+        if self.multDVGeo:
+            DVGeo = self.DVGeos[DVGeoName]
+        else:
+            DVGeo = self.DVGeo
+
         # can only add a child to a FFD DVGeo
         if DVGeo.geoType != "ffd":
             raise RuntimeError(
@@ -146,7 +151,7 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             if pointSet not in DVGeo.children[-1].points:
                 DVGeo.children[-1].addPointSet(DVGeo.points[pointSet], pointSet)
 
-    def nom_add_discipline_coords(self, discipline, DVGeo=None, points=None):
+    def nom_add_discipline_coords(self, discipline, DVGeoName=None, points=None):
         # TODO remove one of these methods to keep only one method to add pointsets
 
         if points is None:
@@ -155,12 +160,24 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             self.add_output("x_%s0" % discipline, distributed=True, copy_shape="x_%s_in" % discipline)
 
         else:
+            # if we have multiple DVGeos use the one specified by name
+            if self.multDVGeo:
+                DVGeo = self.DVGeos[DVGeoName]
+            else:
+                DVGeo = self.DVGeo
+
             # we are provided with points. we can do the full initialization now
             self.nom_addPointSet(DVGeo, points, "x_%s0" % discipline, add_output=False)
             self.add_input("x_%s_in" % discipline, distributed=True, val=points.flatten())
             self.add_output("x_%s0" % discipline, distributed=True, val=points.flatten())
 
-    def nom_addPointSet(self, DVGeo, points, ptName, add_output=True, **kwargs):
+    def nom_addPointSet(self, points, ptName, add_output=True, DVGeoName=None, **kwargs):
+        # if we have multiple DVGeos use the one specified by name
+        if self.multDVGeo:
+            DVGeo = self.DVGeos[DVGeoName]
+        else:
+            DVGeo = self.DVGeo
+
         # add the points to the dvgeo object
         DVGeo.addPointSet(points.reshape(len(points) // 3, 3), ptName, **kwargs)
         self.omPtSetList.append(ptName)
@@ -181,36 +198,49 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         for k, v in point_dict.items():
             self.nom_addPointSet(v, k)
 
-    def nom_getDVGeo(self, name, childIdx=None):
+    def nom_getDVGeo(self, childIdx=None, DVGeoName=None):
         """
-        Gets the DVGeometry object held in the geometry component so DVGeo methods can be called directly on it
+         Gets the DVGeometry object held in the geometry component so DVGeo methods can be called directly on it
+
         Parameters
         ----------
+        DVGeoName : string, optional
+            The name of the DVGeo to return, necessary if there are multiple DVGeo objects
         childIdx : int, optional
-            The zero-based index of the child FFD, you want a child DVGeo returned
+            The zero-based index of the child FFD, if you want a child DVGeo returned
+
[…]
