+++ pygeo/mphys/mphys_dvgeo.py
@@ -246,7 +246,13 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
     Wrapper for DVGeo functions
     """
 
-    def nom_addGlobalDV(self, DVGeo, dvName, value, func, childIdx=None, isComposite=False):
+    def nom_addGlobalDV(self, dvName, value, func, childIdx=None, isComposite=False, DVGeoName=None):
+        # if we have multiple DVGeos use the one specified by name
+        if self.multDVGeo:
+            DVGeo = self.DVGeos[DVGeoName]
+        else:
+            DVGeo = self.DVGeo
+
         # global DVs are only added to FFD-based DVGeo objects
         if DVGeo.geoType != "ffd":
             raise RuntimeError(f"Only FFD-based DVGeo objects can use global DVs, not type:{DVGeo.geoType}")
@@ -264,7 +270,13 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         else:
             DVGeo.children[childIdx].addGlobalDV(dvName, value, func)
 
-    def nom_addLocalDV(self, DVGeo, dvName, axis="y", pointSelect=None, childIdx=None, isComposite=False):
+    def nom_addLocalDV(self, dvName, axis="y", pointSelect=None, childIdx=None, isComposite=False, DVGeoName=None):
+        # if we have multiple DVGeos use the one specified by name
+        if self.multDVGeo:
+            DVGeo = self.DVGeos[DVGeoName]
+        else:
+            DVGeo = self.DVGeo
+
         # local DVs are only added to FFD-based DVGeo objects
         if DVGeo.geoType != "ffd":
             raise RuntimeError(f"Only FFD-based DVGeo objects can use local DVs, not type:{DVGeo.geoType}")
@@ -283,7 +295,6 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
 
     def nom_addLocalSectionDV(
         self,
-        DVGeo,
         dvName,
         secIndex,
         childIdx=None,
@@ -293,7 +304,14 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         orient0=None,
         orient2="svd",
         config=None,
+        DVGeoName=None,
     ):
+        # if we have multiple DVGeos use the one specified by name
+        if self.multDVGeo:
+            DVGeo = self.DVGeos[DVGeoName]
+        else:
+            DVGeo = self.DVGeo
+
         # local DVs are only added to FFD-based DVGeo objects
         if DVGeo.geoType != "ffd":
             raise RuntimeError(f"Only FFD-based DVGeo objects can use local DVs, not type:{DVGeo.geoType}")
@@ -318,7 +336,13 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         self.add_input(dvName, distributed=False, shape=nVal)
         return nVal
 
-    def nom_addShapeFunctionDV(self, DVGeo, dvName, shapes, childIdx=None, config=None):
+    def nom_addShapeFunctionDV(self, dvName, shapes, childIdx=None, config=None, DVGeoName=None):
+        # if we have multiple DVGeos use the one specified by name
+        if self.multDVGeo:
+            DVGeo = self.DVGeos[DVGeoName]
+        else:
+            DVGeo = self.DVGeo
+
         # shape function DVs are only added to FFD-based DVGeo objects
         if DVGeo.geoType != "ffd":
             raise RuntimeError(f"Only FFD-based DVGeo objects can use local DVs, not type:{DVGeo.geoType}")
@@ -334,7 +358,13 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         self.add_input(dvName, distributed=False, shape=nVal)
         return nVal
 
[…]
