+++ pygeo/mphys/orig_mphys.py
@@ -1,0 +1,604 @@
+# External modules
+from mpi4py import MPI
+import numpy as np
+import openmdao.api as om
+from openmdao.api import AnalysisError
+
+# Local modules
+from .. import DVConstraints, DVGeometry, DVGeometryESP, DVGeometryVSP
+
+
+# class that actually calls the dvgeometry methods
+class OM_DVGEOCOMP(om.ExplicitComponent):
+    def initialize(self):
+        self.options.declare("file", default=None)
+        self.options.declare("type", default=None)
+        self.options.declare("options", default=None)
+
+    def setup(self):
+        self.geo_type = self.options["type"]
+
+        # create the DVGeo object that does the computations
+        if self.geo_type == "ffd":
+            # we are doing an FFD-based DVGeo
+            if self.options["options"] is None:
+                ffd_options = {}
+            else:
+                ffd_options = self.options["options"]
+
+            self.DVGeo = DVGeometry(self.options["file"], **ffd_options)
+
+        elif self.geo_type == "vsp":
+            # we are doing a VSP-based DVGeo
+            if self.options["options"] is None:
+                vsp_options = {}
+            else:
+                vsp_options = self.options["options"]
+
+            self.DVGeo = DVGeometryVSP(self.options["file"], comm=self.comm, **vsp_options)
+
+        elif self.geo_type == "esp":
+            # we are doing an ESP-based DVGeo
+            if self.options["options"] is None:
+                esp_options = {}
+            else:
+                esp_options = self.options["options"]
+
+            self.DVGeo = DVGeometryESP(self.options["file"], comm=self.comm, **esp_options)
+
+        self.DVCon = DVConstraints()
+        self.DVCon.setDVGeo(self.DVGeo)
+        self.omPtSetList = []
+
+    def compute(self, inputs, outputs):
+        # check for inputs that have been added but the points have not been added to dvgeo
+        for var in inputs.keys():
+            # check that the input name matches the convention for points
+            if var[:2] == "x_":
+                # trim the _in and add a "0" to signify that these are initial conditions initial
+                var_out = var[:-3] + "0"
+                if var_out not in self.omPtSetList:
+                    self.nom_addPointSet(inputs[var], var_out, add_output=False)
+
+        # inputs are the geometric design variables
+        self.DVGeo.setDesignVars(inputs)
+
+        # ouputs are the coordinates of the pointsets we have
+        for ptName in self.DVGeo.points:
+            if ptName in self.omPtSetList:
+                # update this pointset and write it as output
+                outputs[ptName] = self.DVGeo.update(ptName).flatten()
+
+        # compute the DVCon constraint values
+        constraintfunc = dict()
+        self.DVCon.evalFunctions(constraintfunc, includeLinear=True)
+
+        for constraintname in constraintfunc:
+            # if any constraint returned a fail flag throw an error to OpenMDAO
+            # all constraints need the same fail flag, no <name_> prefix
+            if constraintname == "fail":
+                raise AnalysisError("Analysis error in geometric constraints")
+            outputs[constraintname] = constraintfunc[constraintname]
+
+        # we ran a compute so the inputs changed. update the dvcon jac
+        # next time the jacvec product routine is called
+        self.update_jac = True
+
+    def nom_addChild(self, ffd_file):
+        # can only add a child to a FFD DVGeo
+        if self.geo_type != "ffd":
+            raise RuntimeError(
+                f"Only FFD-based DVGeo objects can have children added to them, not type:{self.geo_type}"
+            )
+
+        # Add child FFD
+        child_ffd = DVGeometry(ffd_file, child=True)
[…]
