+++ pygeo/mphys/mphys_dvgeo.py
@@ -11,7 +11,7 @@ from .. import DVConstraints, DVGeometry, DVGeometryESP, DVGeometryVSP
 # class that actually calls the DVGeometry methods
 class OM_DVGEOCOMP(om.ExplicitComponent):
     def initialize(self):
-        # set up a geometry component with only 1 DVGeo (stadard)
+        # set up a geometry component with only 1 DVGeo (standard)
         self.options.declare("file", default=None)
         self.options.declare("type", default=None)
         self.options.declare("options", default=None)
@@ -20,6 +20,7 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         self.options.declare("DVGeoInfo", default=None)
 
     def setup(self):
+        # create a constraints object
         self.DVCon = DVConstraints()
 
         # create the DVGeo object that does the computations (or multiple DVGeo objects)
@@ -151,7 +152,7 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             if pointSet not in DVGeo.children[-1].points:
                 DVGeo.children[-1].addPointSet(DVGeo.points[pointSet], pointSet)
 
-    def nom_add_discipline_coords(self, discipline, DVGeoName=None, points=None):
+    def nom_add_discipline_coords(self, discipline, points=None, DVGeoName=None):
         # TODO remove one of these methods to keep only one method to add pointsets
 
         if points is None:
@@ -160,14 +161,8 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
             self.add_output("x_%s0" % discipline, distributed=True, copy_shape="x_%s_in" % discipline)
 
         else:
-            # if we have multiple DVGeos use the one specified by name
-            if self.multDVGeo:
-                DVGeo = self.DVGeos[DVGeoName]
-            else:
-                DVGeo = self.DVGeo
-
             # we are provided with points. we can do the full initialization now
-            self.nom_addPointSet(DVGeo, points, "x_%s0" % discipline, add_output=False)
+            self.nom_addPointSet(points, "x_%s0" % discipline, add_output=False, DVGeoName=DVGeoName)
             self.add_input("x_%s_in" % discipline, distributed=True, val=points.flatten())
             self.add_output("x_%s0" % discipline, distributed=True, val=points.flatten())
 
@@ -206,16 +201,15 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
         ----------
         DVGeoName : string, optional
             The name of the DVGeo to return, necessary if there are multiple DVGeo objects
+            
         childIdx : int, optional
             The zero-based index of the child FFD, if you want a child DVGeo returned
 
         Returns
         -------
-        DVGeo object
+       DVGeometry object
+            DVGeometry object held by this geometry component
         """
-        y = 0
-
-
         # if we have multiple DVGeos use the one specified by name
         if self.multDVGeo:
             DVGeo = self.DVGeos[DVGeoName]
@@ -247,6 +241,37 @@ class OM_DVGEOCOMP(om.ExplicitComponent):
     """
 
     def nom_addGlobalDV(self, dvName, value, func, childIdx=None, isComposite=False, DVGeoName=None):
+        """
+        Add a global design variable to the DVGeo object. This is a wrapper for the DVGeo.addGlobalDV method.
+
+        Parameters
+        ----------
+        dvName : str
+            See :meth:`addGlobalDV <.DVGeometry.addGlobalDV>`
+
+        value : float, or iterable list of floats
+            See :meth:`addGlobalDV <.DVGeometry.addGlobalDV>`
+
+        func : python function
+            See :meth:`addGlobalDV <.DVGeometry.addGlobalDV>`
+
+        childIdx : int, optional
+            The zero-based index of the child FFD, if this DV is for a child FFD.
[…]
