### 2023-07-11 d526db7
bring trisurfcon back to original arguments. fix global DV call
