+++ pygeo/constraints/DVCon.py
@@ -1383,7 +1383,7 @@ class DVConstraints:
             This should be the surface with the larger number of triangles.
             By default, it's the ADflow triangulated surface mesh.
 
-        DVGeo1 : DVGeo or None
+        DVGeo_1_name : str or None
             The DVGeo object to associate surface_1 to.
             If None, surface_1 will remain static during optimization.
             By default, it's the 'default' DVGeo object.
@@ -1392,7 +1392,7 @@ class DVConstraints:
             The name of the second triangulated surface to constrain.
             This should be the surface with the smaller number of triangles.
 
-        DVGeo2 : DVGeo or None
+        DVGeo_1_name : str or None
             The DVGeo object to associate surface_2 to.
             If None, surface_2 will remain static during optimization.
             By default, it's the 'default' DVGeo object.
