### 2023-07-12 02f46b1
fillet intersection class
