+++ pygeo/parameterization/DVGeoMulti.py
@@ -114,7 +114,13 @@ class DVGeometryMulti:
 
         if self.fillet is False and points is not None:
             raise Error("Unstructured point data is only valid for fillet DVGeoMulti")
-        
+
+        # determine whether this component is a fillet or a normal surface
+        if DVGeo is None:
+            filletComp = True
+        else:
+            filletComp = False
+
         # if self.fillet is True and triMesh is not None:
         # this should work with a triangulated surface it just isn't necessary
 
@@ -135,6 +141,11 @@ class DVGeometryMulti:
             # add these points to the corresponding dvgeo
             if DVGeo is not None:
                 DVGeo.addPointSet(nodes, "datPts", **pointSetKwargs)
+
+            nodes = surfPts
+            triConn = None
+            triConnStack = None
+            barsConn = None
 
         # we have a standard intersection group which has structured surfaces
         else:
@@ -172,7 +183,7 @@ class DVGeometryMulti:
             xMax[2] = bbox["zmax"]
 
         # initialize the component object
-        self.comps[comp] = component(comp, DVGeo, nodes, triConn, triConnStack, barsConn, xMin, xMax)
+        self.comps[comp] = component(comp, DVGeo, nodes, filletComp, triConn, triConnStack, barsConn, xMin, xMax)
 
         # add the name to the list
         self.compNames.append(comp)
@@ -184,6 +195,7 @@ class DVGeometryMulti:
         self,
         compA,
         compB,
+        filletComp=None,
         dStarA=0.2,
         dStarB=0.2,
         featureCurves=None,
@@ -274,6 +286,9 @@ class DVGeometryMulti:
 
         """
 
+        if filletComp is not None and not self.fillet:
+            print("no")
+
         # Assign mutable defaults
         if featureCurves is None:
             featureCurves = []
@@ -289,6 +304,7 @@ class DVGeometryMulti:
             CompIntersection(
                 compA,
                 compB,
+                filletComp,
                 dStarA,
                 dStarB,
                 featureCurves,
@@ -328,7 +344,7 @@ class DVGeometryMulti:
             or when getting the derivatives of the coordinates.
         compNames : list, optional
             A list of component names that this point set should be added to.
-            To ease bookkeepping, an empty point set with ptName will be added to components not in this list.
+            To ease bookkeeping, an empty point set with ptName will be added to components not in this list.
             If a list is not provided, this point set is added to all components.
         comm : MPI.IntraComm, optional
             Comm that is associated with the added point set. Does not
@@ -961,10 +977,11 @@ class DVGeometryMulti:
 
 
 class component:
-    def __init__(self, name, DVGeo, nodes, triConn, triConnStack, barsConn, xMin, xMax):
+    def __init__(self, name, DVGeo, fillet, nodes, triConn, triConnStack, barsConn, xMin, xMax):
         # save the info
         self.name = name
         self.DVGeo = DVGeo
+        self.fillet = fillet
         self.nodes = nodes
         self.triConn = triConn
         self.triConnStack = triConnStack
@@ -982,23 +999,14 @@ class component:
             self.triMesh = True
 
     def updateTriMesh(self):
+        if self.fillet:
+            pointset = self.name
+        else:
+            pointset = "trimesh"
+
         # update the triangulated surface mesh
-        self.nodes = self.DVGeo.update("triMesh")
+        self.nodes = self.DVGeo.update(pointset)
 
-
-class filletComp:
-    def __init__(self, name, nodes, xMin, xMax, DVGeo=None):
-        self.name = name
[…]
