### 2023-07-14 3d68595
Merge `_getSurfacePoints` into `_getSurfaceVertices`
