+++ pygeo/constraints/DVCon.py
@@ -374,7 +374,7 @@ class DVConstraints:
             Name of the DVGeo object to obtain the surface from (default is 'None')
         """
 
-        p0, p1, p2 = self._getSurfacePoints(surfaceName, fromDVGeo)
+        p0, p1, p2 = self._getSurfaceVertices(surfaceName, fromDVGeo)
 
         f = open(fileName, "w")
         f.write('TITLE = "DVConstraints Surface Mesh"\n')
@@ -411,7 +411,7 @@ class DVConstraints:
         except ImportError as e:
             raise ImportError("numpy-stl package must be installed") from e
 
-        p0, p1, p2 = self._getSurfacePoints(surfaceName, fromDVGeo)
+        p0, p1, p2 = self._getSurfaceVertices(surfaceName, fromDVGeo)
 
         stlmesh = mesh.Mesh(np.zeros(p0.shape[0], dtype=mesh.Mesh.dtype))
         stlmesh.vectors[:, 0, :] = p0
@@ -420,30 +420,6 @@ class DVConstraints:
 
         # Write the mesh to file "cube.stl"
         stlmesh.save(fileName)
-
-    def _getSurfacePoints(self, surfaceName="default", fromDVGeo=None):
-        """Get the points that define a triangulated surface mesh.
-
-        Parameters
-        ----------
-        surfaceName : str
-            Which DVConstraints surface to get the points for (default is 'default')
-        fromDVGeo : str or None
-            Name of the DVGeo object to obtain the surface from (default is 'None' in which case the surface is obtained
-            from the DVConstraints object itself)
-
-        Returns
-        -------
-        (np.array, np.array, np.array)
-            Arrays of points that define the triangulated surface mesh
-        """
-        if fromDVGeo is None:
-            p0, p1, p2 = self._getSurfaceVertices(surfaceName=surfaceName)
-        else:
-            p0 = self.DVGeometries[fromDVGeo].update(surfaceName + "_p0")
-            p1 = self.DVGeometries[fromDVGeo].update(surfaceName + "_p1")
-            p2 = self.DVGeometries[fromDVGeo].update(surfaceName + "_p2")
-        return p0, p1, p2
 
     def addThicknessConstraints2D(
         self,
@@ -3237,12 +3213,32 @@ class DVConstraints:
                 "constraints can be added."
             )
 
-    def _getSurfaceVertices(self, surfaceName):
-        if surfaceName not in self.surfaces.keys():
-            raise KeyError('Need to add surface "' + surfaceName + '" to the DVConstraints object')
-        p0 = self.surfaces[surfaceName][0]
-        p1 = self.surfaces[surfaceName][1]
-        p2 = self.surfaces[surfaceName][2]
+    def _getSurfaceVertices(self, surfaceName="default", fromDVGeo=None):
+        """Get the points that define a triangulated surface mesh.
+
+        Parameters
+        ----------
+        surfaceName : str
+            Which DVConstraints surface to get the points for (default is 'default')
+        fromDVGeo : str or None
+            Name of the DVGeo object to obtain the surface from (default is 'None' in which case the surface is obtained
+            from the DVConstraints object itself)
+
+        Returns
+        -------
+        (np.array, np.array, np.array)
+            Arrays of points that define the triangulated surface mesh
+        """
+        if fromDVGeo is None:
+            if surfaceName not in self.surfaces.keys():
+                raise KeyError('Need to add surface "' + surfaceName + '" to the DVConstraints object')
+            p0 = self.surfaces[surfaceName][0]
+            p1 = self.surfaces[surfaceName][1]
+            p2 = self.surfaces[surfaceName][2]
+        else:
[…]
