### 2023-07-14 0730b2b
Add explanation for `fromDVGeo=None` case
