### 2023-07-17 de245fb
fillet interface for intersection curve
