+++ pygeo/constraints/DVCon.py
@@ -1385,7 +1385,7 @@ class DVConstraints:
             By default, it's the ADflow triangulated surface mesh.
 
         DVGeo_1_name : str or None
-            The DVGeo object to associate surface_1 to.
+            The name of the DVGeo object to associate surface_1 to.
             If None, surface_1 will remain static during optimization.
             By default, it's the 'default' DVGeo object.
 
@@ -1394,7 +1394,7 @@ class DVConstraints:
             This should be the surface with the smaller number of triangles.
 
         DVGeo_1_name : str or None
-            The DVGeo object to associate surface_2 to.
+            The name of the DVGeo object to associate surface_2 to.
             If None, surface_2 will remain static during optimization.
             By default, it's the 'default' DVGeo object.
 
