### 2023-07-17 8642ed3
remove string types
