+++ pygeo/parameterization/DVGeoSketch.py
@@ -65,7 +65,7 @@ class DVGeoSketch(BaseDVGeometry):
     """
 
     def __init__(self, fileName, geoType, comm=MPI.COMM_WORLD, scale=1.0, projTol=0.01, name=None):
-        super().__init__(fileName=fileName, name=name, geoType=geoType)
+        super().__init__(fileName=fileName, name=name)
 
         # this scales coordinates from model to mesh geometry
         self.modelScale = scale
