### 2023-07-17 a046415
save a few lines
