### 2023-07-17 56c89ce
can add curves to components in intersection
