+++ pygeo/parameterization/DVGeoMulti.py
@@ -146,6 +146,7 @@ class DVGeometryMulti:
             triConn = None
             triConnStack = None
             barsConn = None
+            component = Comp(comp, filletComp, nodes, DVGeo)
 
         # we have a standard intersection group which has structured surfaces
         else:
@@ -165,8 +166,6 @@ class DVGeometryMulti:
                 triConnStack = None
                 barsConn = None
 
-        # get bounding box information if this component has a DVGeo (is not a fillet)
-        if DVGeo is not None:
             # we will need the bounding box information later on, so save this here
             xMin, xMax = DVGeo.FFD.getBounds()
 
@@ -184,13 +183,10 @@ class DVGeometryMulti:
             if "zmax" in bbox:
                 xMax[2] = bbox["zmax"]
 
-        # dummy bounding box for fillet component
-        else:
-            xMin = 3 * [0]
-            xMax = 3 * [0]
+            component = component(comp, DVGeo, nodes, filletComp, triConn, triConnStack, barsConn, xMin, xMax)
 
         # initialize the component object
-        self.comps[comp] = component(comp, DVGeo, nodes, filletComp, triConn, triConnStack, barsConn, xMin, xMax)
+        self.comps[comp] = component
 
         # add the name to the list
         self.compNames.append(comp)
@@ -298,7 +294,7 @@ class DVGeometryMulti:
             if filletComp is None:
                 print("no")
 
-            inter = FilletIntersection(compA, compB, filletComp, self, self.dtype)
+            inter = FilletIntersection(compA, compB, filletComp, distTol, self, self.dtype)
 
         # initialize a standard intersection object
         else:
@@ -315,7 +311,6 @@ class DVGeometryMulti:
             inter = CompIntersection(
                 compA,
                 compB,
-                filletComp,
                 dStarA,
                 dStarB,
                 featureCurves,
@@ -344,10 +339,10 @@ class DVGeometryMulti:
 
         comp = self.comps[compName]
         fillet = self.comps["fillet"]
-        intersection = self.intersectComps[fillet.intersection]
+        intersection = fillet.intersection
 
-        filletIntCurve, filletIntInd = intersection.findIntersection(fillet.pts, curvePts)
-        compIntCurve, compIntInd = intersection.findIntersection(comp.pts, curvePts)
+        filletIntCurve, filletIntInd = intersection.findIntersection(fillet.surfPts, curvePts)
+        compIntCurve, compIntInd = intersection.findIntersection(comp.surfPts, curvePts)
 
         fillet.intCurve = filletIntCurve
         fillet.intInd = filletIntInd
@@ -1045,8 +1040,8 @@ class Comp:
         self.DVGeo = DVGeo
         self.surfPts = surfPts
         self.surfPtsOrig = deepcopy(surfPts)
-        self.curvePts = curvePts
 
+        self.intersection = None
         self.intersectPts = []
         self.intersectInd = []
         self.adjacentComponents = []
@@ -1062,15 +1057,18 @@ class PointSet:
 
 
 class Intersection:
-    def __init__(self, dtype, compA, compB, DVGeo):
-        self.dtype = dtype
-
+    def __init__(self, compA, compB, distTol, DVGeo, dtype):
         componentA = DVGeo.comps[compA]
         componentB = DVGeo.comps[compB]
 
         self.compA = componentA
         self.compB = componentB
         self.DVGeo = DVGeo
+        self.dtype = dtype
+        self.distTol = distTol
+
+        # same communicator with DVGeo
+        self.comm = DVGeo.comm
 
     def setSurface(self, comm):
         """This set the new udpated surface on which we need to compute the new intersection curve"""
[…]
