### 2023-07-17 f9ce47f
distinguish between intersection curves on fillet
