### 2023-07-18 2da39cd
add pointset to multi for fillet intersection component
