+++ pygeo/parameterization/DVGeoMulti.py
@@ -8,6 +8,8 @@ import numpy as np
 from scipy import sparse
 from scipy.spatial.distance import cdist
 from copy import deepcopy
+from pyspline.utils import closeTecplot, openTecplot, writeTecplot1D
+
 
 try:
     # External modules
@@ -146,7 +148,6 @@ class DVGeometryMulti:
             triConn = None
             triConnStack = None
             barsConn = None
-            component = Comp(comp, filletComp, nodes, DVGeo)
 
         # we have a standard intersection group which has structured surfaces
         else:
@@ -166,6 +167,7 @@ class DVGeometryMulti:
                 triConnStack = None
                 barsConn = None
 
+        if not filletComp:
             # we will need the bounding box information later on, so save this here
             xMin, xMax = DVGeo.FFD.getBounds()
 
@@ -182,8 +184,13 @@ class DVGeometryMulti:
                 xMax[1] = bbox["ymax"]
             if "zmax" in bbox:
                 xMax[2] = bbox["zmax"]
+        else:
+            xMin = xMax = 3 * [0]
 
-            component = component(comp, DVGeo, nodes, filletComp, triConn, triConnStack, barsConn, xMin, xMax)
+        if self.fillet:
+            component = Comp(comp, filletComp, nodes, DVGeo, xMin, xMax)
+        else:
+            component = component(comp, DVGeo, nodes, triConn, triConnStack, barsConn, xMin, xMax)
 
         # initialize the component object
         self.comps[comp] = component
@@ -387,35 +394,36 @@ class DVGeometryMulti:
 
         # before we do anything, we need to create surface ADTs
         # for which the user provided triangulated meshes
-        for comp in compNames:
-            # check if we have a trimesh for this component
-            if self.comps[comp].triMesh:
-                # Now we build the ADT using pySurf
-                # Set bounding box for new tree
-                BBox = np.zeros((2, 3))
-                useBBox = False
+        if not self.fillet:
+            for comp in compNames:
+                # check if we have a trimesh for this component
+                if self.comps[comp].triMesh:
+                    # Now we build the ADT using pySurf
+                    # Set bounding box for new tree
+                    BBox = np.zeros((2, 3))
+                    useBBox = False
 
-                # dummy connectivity data for quad elements since we have all tris
-                quadConn = np.zeros((0, 4))
+                    # dummy connectivity data for quad elements since we have all tris
+                    quadConn = np.zeros((0, 4))
 
-                # Compute set of nodal normals by taking the average normal of all
-                # elements surrounding the node. This allows the meshing algorithms,
-                # for instance, to march in an average direction near kinks.
-                nodal_normals = self.adtAPI.adtcomputenodalnormals(
-                    self.comps[comp].nodes.T, self.comps[comp].triConnStack.T, quadConn.T
-                )
-                self.comps[comp].nodal_normals = nodal_normals.T
+                    # Compute set of nodal normals by taking the average normal of all
+                    # elements surrounding the node. This allows the meshing algorithms,
+                    # for instance, to march in an average direction near kinks.
+                    nodal_normals = self.adtAPI.adtcomputenodalnormals(
+                        self.comps[comp].nodes.T, self.comps[comp].triConnStack.T, quadConn.T
+                    )
+                    self.comps[comp].nodal_normals = nodal_normals.T
 
-                # Create new tree (the tree itself is stored in Fortran level)
-                self.adtAPI.adtbuildsurfaceadt(
-                    self.comps[comp].nodes.T,
[…]
