+++ pygeo/parameterization/DVGeoMulti.py
@@ -52,20 +52,23 @@ class DVGeometryMulti:
 
     """
 
-    def __init__(self, comm=MPI.COMM_WORLD, fillet=False, checkDVs=True, debug=False, isComplex=False):
+    def __init__(self, comm=MPI.COMM_WORLD, filletIntersection=False, checkDVs=True, debug=False, isComplex=False):
         # Check to make sure pySurf is installed before initializing
-        if not pysurfInstalled and not fillet:
+        if not pysurfInstalled and not filletIntersection:
             raise ImportError("pySurf is not installed and is required to use DVGeometryMulti outside of fillet mode.")
 
         self.compNames = []
         self.comps = OrderedDict()
+        # separate out fillets to avoid checking whether components in comps have a DVGeo everywhere
+        self.filletNames = []  
+        self.fillets = OrderedDict()
         self.DVGeoDict = OrderedDict()
         self.points = OrderedDict()
         self.updated = {}
         self.intersectComps = []
 
         self.comm = comm
-        self.fillet = fillet
+        self.filletIntersection = filletIntersection
         self.checkDVs = checkDVs
         self.debug = debug
         self.complex = isComplex
@@ -111,10 +114,10 @@ class DVGeometryMulti:
             Keyword arguments to be passed to the component addPointSet call for the triangulated mesh.
 
         """
-        if DVGeo is None and self.fillet is False:
+        if DVGeo is None and self.filletIntersection is False:
             raise Error("DVGeo must be assigned for non-fillet DVGeoMulti components")
 
-        if self.fillet is False and points is not None:
+        if self.filletIntersection is False and points is not None:
             raise Error("Unstructured point data is only valid for fillet DVGeoMulti")
 
         # determine whether this component is a fillet or a normal surface
@@ -123,7 +126,7 @@ class DVGeometryMulti:
         else:
             filletComp = False
 
-        # if self.fillet is True and triMesh is not None:
+        # if self.filletIntersection is True and triMesh is not None:
         # this should work with a triangulated surface it just isn't necessary
 
         # Assign mutable defaults
@@ -132,8 +135,31 @@ class DVGeometryMulti:
         if pointSetKwargs is None:
             pointSetKwargs = {}
 
+        # fillets don't have a DVGeo to get a bounding box from and don't need it
+        if filletComp:
+            xMin = xMax = 3 * [0]
+
+        # standard components need a bounding box to associate points with each FFD
+        else:
+            # we will need the bounding box information later on, so save this here
+            xMin, xMax = DVGeo.FFD.getBounds()
+
+            # also we might want to modify the bounding box if the user specified any coordinates
+            if "xmin" in bbox:
+                xMin[0] = bbox["xmin"]
+            if "ymin" in bbox:
+                xMin[1] = bbox["ymin"]
+            if "zmin" in bbox:
+                xMin[2] = bbox["zmin"]
+            if "xmax" in bbox:
+                xMax[0] = bbox["xmax"]
+            if "ymax" in bbox:
+                xMax[1] = bbox["ymax"]
+            if "zmax" in bbox:
+                xMax[2] = bbox["zmax"]
+
         # we have a fillet so no structured surfaces are necessary
-        if self.fillet:
+        if self.filletIntersection:
             # save unstructured point data
             surfPts = self._readDATFile(points, surf=True)
 
@@ -141,13 +167,13 @@ class DVGeometryMulti:
             surfPts *= scale
             nodes = surfPts
 
-            # add these points to the corresponding dvgeo
-            if DVGeo is not None:
+            # add these points to the corresponding dvgeo unless this component is a fillet
+            if not filletComp:
[…]
