### 2023-07-19 #5319 @danielskatz
@hajdik - I'm now proofreading this paper to get ready to process it further towards publication. While I do so, I notice that the zenodo archive doesn't have any ORCIDs for the authors, while the paper does. If you want to, you can edit the Zenodo metadata to add ORCIDs. This is up to you to decide.
